# [Code] Remove Regex Compile Bloat