--- a/zone/cli/tests/npc_handins_multiquest.cpp
+++ b/zone/cli/tests/npc_handins_multiquest.cpp
@@ -1,4 +1,3 @@
-#include "../../common/http/httplib.h"
 #include "../../common/eqemu_logsys.h"
 #include "../../common/platform.h"
 #include "../../zone.h"